requirements.txt
# DeepFake Detector - Dependencies
# Python 3.9+

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
face-recognition>=1.3.0
Pillow>=10.0.0

# ML Utilities
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.65.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Utils
pydantic>=2.0.0
aiofiles>=23.0.0